refactor : 상품공통정보 - 기존정보목록 컴포넌트 분리
